Added delete, +/- buttons for cart items. Attempt to add a grid for the cart. Added TODOs for CART.

+++ src/components/Cart.vue
@@ -1,15 +1,35 @@
 <template>
     <div>
         <h1> Your Cart items</h1>
-        <div v-for="(order,key) in orderToCart" :key="key">
-            {{"Burger " +key}} : {{order}}
+        <!--TODO: Fix grid below for the cart!-->
+        <!--TODO: Show only the grid when the cart is not empty!-->
+        <div class="ing-grid">
+            <div>Item</div>
+            <div>Ingredients</div>
+            <div>Unit Price</div>
+            <div>Quantity</div>
+            <div>Total</div>
+            <div>Delete Item</div>
         </div>
+        <div v-for="(order,key) in orderToCart" :key="key" class="ing-grid">
+            <div>{{key}}</div>
+            <div>{{order.order.ingredients.map(item=>item["ingredient_"+ lang]).join(", ") }}</div>
+            <div>{{order.order.price}}</div>
+            <div>
+                <button v-on:click="decrementQuantity(key)"> - </button>
+                {{order.order.quantity}}
+                <button v-on:click="incrementQuantity(key)"> + </button>
+            </div>
+            <div>{{(order.order.price * order.order.quantity)}}</div>
+            <button v-on:click="deleteItem(key)">Delete</button>
+        </div>
+        <br>
         <div v-if="totalPrice != '0'">
             Total Price: {{totalPrice}} SEK
         </div>
         <div>
-            <button v-on:click="placeOrder()">Order</button>
             <button v-on:click="clearCart()">Clear Cart</button>
+            <button v-on:click="placeOrder()">Order</button>
             <h1 v-if="orderNumber !== ''">  {{"your order number is: " + orderNumber}}</h1>
             <h1>{{ uiLabels.ordersInQueue }}</h1>
             <div v-for="(order, orderkey) in orders"
@@ -42,7 +62,8 @@
                 orderToCart:{},
                 currentItemsCount:0,
                 totalPrice:0,
-                orderNumber:""
+                orderNumber:"",
+                selectedItems:[],
             }
 
         },
@@ -91,11 +112,45 @@
                 this.currentItemsCount += 1;
                 return this.currentItemsCount;
             },
+            getSelectedItems: function(){
+                var items,i, item_key;
+                items={};
+                for (i = 0; i < this.selectedItems.length; i += 1) {
+                    item_key = this.selectedItems[i];
+                    items[item_key]=this.orderToCart[item_key];
+                }
+                return items;
+            },
+            incrementQuantity: function(itemKey){
+                if (this.orderToCart[itemKey].order.quantity  < this.orderToCart[itemKey].order.stock){
+                    this.orderToCart[itemKey].order.quantity += 1;
+                    this.totalPrice += this.orderToCart[itemKey].order.price;
+                }
+                else{
+                    alert("Maximum number of Item "+ itemKey +" reached! Not enough in stock.")
+                }
+            },
+            decrementQuantity: function(itemKey){
+                if (this.orderToCart[itemKey].order.quantity > 1)
+                    {
+                        this.orderToCart[itemKey].order.quantity -= 1;
+                        this.totalPrice -= this.orderToCart[itemKey].order.price;
+                    }
+
+            },
+            deleteItem: function(itemKey){
+                this.totalPrice -= this.orderToCart[itemKey].order.price;
+                this.currentItemsCount -= 1;
+                delete this.orderToCart[itemKey];
+            },
+
             clearCart: function () {
-                this.currentItemsCount =0;
-                this.totalPrice = 0;
-                this.orderToCart={};
-                localStorage.clear();
+                if (confirm("Are you sure you want to clear your cart?")){
+                    this.currentItemsCount =0;
+                    this.totalPrice = 0;
+                    this.orderToCart={};
+                    localStorage.clear();
+                }
             }
 
         }
@@ -105,5 +160,11 @@
 </script>
 
 <style scoped>
+.ing-grid{
+    display: grid;
+    grid-gap: 10px;
+    grid-template-columns: auto auto auto auto auto auto;
+    text-align: left;
 
+}
 </style>
